requirements.txt
Flask==3.*
Flask-SQLAlchemy==3.*
Flask-Cors==4.*
SQLAlchemy==2.*
PyMySQL==1.1.*
python-dotenv==1.0.*
PyJWT==2.*
